Specifying assets https://flutter.dev/docs/development/ui/assets-and-images#specifying-assets
Flutter uses the pubspec.yaml file, located at the root of your project, to identify assets required by an app.

The build process supports the notion of asset variants: different versions of an asset that might be displayed in different contexts. When an asset’s path is specified in the assets section of pubspec.yaml, the build process looks for any files with the same name in adjacent subdirectories. Such files are then included in the asset bundle along with the specified asset.