Ephemeral state https://flutter.dev/docs/development/data-and-backend/state-mgmt/ephemeral-vs-app#ephemeral-state
Ephemeral state (sometimes called UI state or local state) is the state you can neatly contain in a single widget.

This is, intentionally, a vague definition, so here are a few examples.
-current page in a PageView
-current progress of a complex animation
-current selected tab in a BottomNavigationBar

Other parts of the widget tree seldom need to access this kind of state. There is no need to serialize it, and it doesn’t change in complex ways.

In other words, there is no need to use state management techniques (ScopedModel, Redux, etc.) on this kind of state. All you need is a StatefulWidget